# Postman_collections/s3_ivanysusbambam-web/target/jaxrs-analyzer/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0",
        "title": "ivanysusbambam-web"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/calificacionCarro": {
            "post": {
                "description": "POST /api/ventas/(id)/prospectosCompra: Crea una nueva calificación de carro.\n <pre> \n Codigos de respuesta:\n <code style=\"color: mediumseagreen; background-color: #eaffe0;\">\n 200 OK se creóla nueva calificacion.\n </code> \n <code style=\"color: #c7254e; background-color: #f9f2f4;\">\n 404 Not Found no existe la venta con el id dado.\n </code> \n </pre>",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "id de la venta."
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CalificacionCarroDetailDTO"
                        },
                        "description": "la calificacion que se añadira que se añadirá."
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "headers": {
                        },
                        "schema": {
                            "$ref": "#/definitions/CalificacionCarroDetailDTO"
                        }
                    }
                }
            },
            "delete": {
                "description": "DELETE /api/ventas/(id)/calificacionCarro\n <pre> \n Codigos de respuesta:\n <code style=\"color: mediumseagreen; background-color: #eaffe0;\">\n 200 OK se eliminó la calificacion del carro.\n </code> \n <code style=\"color: #c7254e; background-color: #f9f2f4;\">\n 404 Not Found no existe la venta id dado.\n </code> \n </pre>",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "id del cliente."
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "headers": {
                        },
                        "schema": {
                            "$ref": "#/definitions/CalificacionCarroDetailDTO"
                        }
                    }
                }
            }
        },
...
        "/calificacionesTienda": {
            "get": {
                "description": "GET api/clientes/(id)/calificacionTienda Retorna todas las calificaciones  \n que un cliente ha puesto a una lista de tiendas\n \n <pre>Busca y devuelve todas las calificaciones que un cliente ha dado.\n \n Codigos de respuesta:\n <code style=\"color: mediumseagreen; background-color: #eaffe0;\">\n 200 OK Devuelve todas las calificaciones del cliente.</code> \n </pre>",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "identificador unico del cliente."
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "headers": {
                        },
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/CalificacionTiendaDTO"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "POST /api/clientes/(id)/calificacionTienda: Crea un nuevo prospecto de compra.\n <pre> \n Codigos de respuesta:\n <code style=\"color: mediumseagreen; background-color: #eaffe0;\">\n 200 OK se creó la nueva calificacion.\n </code> \n <code style=\"color: #c7254e; background-color: #f9f2f4;\">\n 404 Not Found no existe el cliente el id dado.\n </code> \n </pre>",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "id del cliente."
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CalificacionTiendaDetailDTO"
                        },
                        "description": "calificacion de la tienda que se añadirá."
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "headers": {
                        },
                        "schema": {
                            "$ref": "#/definitions/CalificacionTiendaDetailDTO"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "CalificacionCarroDetailDTO": {
            "properties": {
                "comentario": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "puntaje": {
                    "type": "integer"
                },
                "venta": {
                    "$ref": "#/definitions/VentaDTO"
                }
            }
        },
        "CalificacionTiendaDTO": {
            "properties": {
                "comentario": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "puntaje": {
                    "type": "integer"
                }
            }
        },
        "CalificacionTiendaDetailDTO": {
            "properties": {
                "cliente": {
                    "$ref": "#/definitions/ClienteDTO"
                },
                "comentario": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "puntaje": {
                    "type": "integer"
                }
            }
        },
        "ClienteDTO": {
            "properties": {
                "cedula": {
                    "type": "integer"
                },
                "nombre": {
                    "type": "string"
                }
            }
        },
...
        "VentaDTO": {
            "properties": {
                "idVenta": {
                    "type": "integer"
                }
            }
        },
...
    }
}